if run from dfm clone, try to find dotfiles repo
The t directory was renamed to dfm_tests because its presence next to
the dfm script is an indicator that dfm is being run from a dfm clone
and not a dotfiles clone.  There is a very low likelyhood that someone
would have a directory called dfm_tests in their dotfiles repo.